Organized the java classes to the model and controller
