+++ E-Mart/nbproject/project.properties
@@ -28,7 +28,6 @@ dist.war=${dist.dir}/${war.name}
 endorsed.classpath=
 excludes=
 file.reference.mysql-connector-j-8.3.0.jar=C:\\Users\\Esanki Lakvindee\\Desktop\\mysql-connector-j-8.3.0\\mysql-connector-j-8.3.0.jar
-file.reference.mysql-connector-j-8.3.0.jar-1=D:\\NetBeans-20\\mysql-connector-j-8.3.0\\mysql-connector-j-8.3.0.jar
 includes=**
 j2ee.compile.on.save=false
 j2ee.copy.static.files.on.save=true
@@ -39,8 +38,7 @@ j2ee.server.type=Tomcat
 jar.compress=false
 javac.classpath=\
     ${libs.jakartaee-api-10.0.classpath}:\
-    ${file.reference.mysql-connector-j-8.3.0.jar}:\
-    ${file.reference.mysql-connector-j-8.3.0.jar-1}
+    ${file.reference.mysql-connector-j-8.3.0.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.debug=true
