+++ E-Mart/nbproject/project.xml
@@ -14,10 +14,6 @@
                     <file>${file.reference.mysql-connector-j-8.3.0.jar}</file>
                     <path-in-war>WEB-INF/lib</path-in-war>
                 </library>
-                <library dirs="200">
-                    <file>${file.reference.mysql-connector-j-8.3.0.jar-1}</file>
-                    <path-in-war>WEB-INF/lib</path-in-war>
-                </library>
             </web-module-libraries>
             <web-module-additional-libraries/>
             <source-roots>
